+++ TADS_TP/DTOs/AluguelResponseDTO.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public DateTime DataInicio { get; set; }
         public DateTime DataFim { get; set; }
-        public decimal ValorTotal { get; set; }
+        public double ValorTotal { get; set; }
 
         public string ClienteNome { get; set; }
         public string VeiculoModelo { get; set; }
