+++ TADS_TP/Models/AluguelModel.cs
@@ -10,8 +10,8 @@
         public double QuilometragemInicial { get; set; }
         public double QuilometragemFinal { get; set; }
 
-        public decimal ValorDiaria { get; set; }
-        public decimal ValorTotal { get; set; }
+        public double ValorDiaria { get; set; }
+        public double ValorTotal { get; set; }
 
         public DateTime? DataDevolucao { get; set; }
